add polices

+++ app/Http/Controllers/Admin/PostController.php
@@ -10,6 +10,8 @@ use App\Post;
 use App\Category;
 use App\Tag;
 
+use Illuminate\Support\Facades\Storage;
+
 class PostController extends Controller
 {
     public function __construct()
@@ -54,6 +56,16 @@ class PostController extends Controller
 
         $post = Post::create($request->all());
 
+        //Imagen
+        if($request->file('file'))
+        {
+            $path = Storage::disk('public')->put('image', $request->file('file'));
+            //dd($path);
+            $post->fill(['file' => asset($path)])->save();
+        }
+        //Tags
+        $post->tags()->sync($request->get('tags'));
+
         return redirect()->route('posts.edit', $post->id)
             ->with('info', 'Entrada creada con exito');
     }
@@ -67,6 +79,7 @@ class PostController extends Controller
     public function show($id)
     {
         $post = Post::find($id);
+        $this->authorize('pass', $post);
 
         return view('admin.posts.show', compact('post'));
     }
@@ -79,9 +92,11 @@ class PostController extends Controller
      */
     public function edit($id)
     {
+        $post = Post::find($id);
+        $this->authorize('pass', $post);
+
         $categories = Category::orderBy('name', 'ASC')->pluck('name', 'id');
         $tags = Tag::orderBy('name', 'ASC')->get();
-        $post = Post::find($id);
 
         return view('admin.posts.edit', compact('post', 'categories', 'tags'));
     }
@@ -96,8 +111,19 @@ class PostController extends Controller
     public function update(PostUpdateRequest $request, $id)
     {
         $post = Post::find($id);
+        $this->authorize('pass', $post);
+
 
         $post->fill($request->all())->save();
+
+        //Imagen
+        if($request->file('file'))
+        {
+            $path = Storage::disk('public')->put('image', $request->file('file'));
+            $post->fill(['file' => asset($path)])->save();
+        }
+        //Tags
+        $post->tags()->sync($request->get('tags'));
 
         return redirect()->route('posts.edit', $post->id)
             ->with('info', 'Entrada actualizada con exito');
@@ -112,6 +138,8 @@ class PostController extends Controller
     public function destroy($id)
     {
         $post = Post::find($id);
+        $this->authorize('pass', $post);
+
 
         $post->delete();
